Refactor issue and team imports and remove debug logs
Update imports for Issue and Team in various modules to use consistent paths. Also, remove unnecessary console log statements to clean up the code.
