+++ src/lib/hooks/issues/issues-context.tsx
@@ -1,13 +1,13 @@
-import { Issue } from '@linear/sdk'
 import { useParams } from 'next/navigation'
 import React, { createContext, ReactNode, useContext, useEffect, useState } from 'react'
 
+import { Issue } from '@aces/interfaces/issue'
 import { getIssuesForView } from '@aces/lib/api/issues/get-issues-for-view'
 import useCurrentUser from '@aces/lib/hooks/auth/use-current-user'
 import { fetchIssuesForTeam } from '@aces/lib/hooks/issues/get-issues-for-team'
 import { setRoundIssue } from '@aces/lib/hooks/rounds/use-set-round-issue'
+import useTeams from '@aces/lib/hooks/teams/teams-context'
 import useViews from '@aces/lib/hooks/views/views-context'
-import useTeams from '@aces/lib/teams/teams-context'
 
 
 interface IssuesContextProps {
@@ -30,12 +30,10 @@ const IssuesProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
   const { selectedTeam } = useTeams()
   const { user } = useCurrentUser()
 
-
   async function loadIssuesForTeams() {
     setIsLoading(true)
     try {
       const { issues } = await fetchIssuesForTeam(selectedTeam!.id)
-      console.log('issues:', issues)
       setIssues(issues)
       const firstIssue: Issue | null = issues[0] || null
       setCurrentIssue(firstIssue)
@@ -56,7 +54,6 @@ const IssuesProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
       loadIssuesForViews()
     }
     if (!selectedView && selectedTeam) {
-      console.log('loading issues for teams')
       loadIssuesForTeams()
     }
     else {
